[FIX] Warning message (Analog channel gain's value) when reading a C3D file exported in C3D and coming from an acquisition stored in a ANC file. [FIX] Rounding error for the number of frames in ANC file under MS Windows (error since the previous patch to fix the same problem under Linux).
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@287 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/IO/btkANBFileIO.cpp
@@ -373,7 +373,7 @@ namespace btk
         if ((*it)->GetGain() == Analog::Unknown)
         {
           channelRange[i] = ANxFileIODetectAnalogRange((*it)->GetScale(), input->GetAnalogResolution());
-          btkErrorMacro("Unknown gain for channel #" + ToString(i+1) + ". Automatically replaced by +/- " + ToString(channelRange[i] / 1000)  + " volts in the file.");
+          btkErrorMacro("Unknown gain for channel #" + ToString(i+1) + ". Automatically replaced by +/- " + ToString(static_cast<double>(channelRange[i]) / 1000.0)  + " volts in the file.");
         }
         else
           channelRange[i] = (*it)->GetGain();
